requirements.txt
appthreat-vulndb
tabulate
defusedxml
PyYAML
